+++ src/main/java/io/github/astrapi69/entity/creatable/ByCreation.java
@@ -29,6 +29,7 @@ import java.io.Serializable;
 import javax.persistence.Entity;
 import javax.persistence.MappedSuperclass;
 
+import io.github.astrapi69.data.creatable.IdentifiableByCreatable;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -56,8 +57,7 @@ import io.github.astrapi69.entity.base.SequenceBaseEntity;
 @NoArgsConstructor
 @SuperBuilder
 public abstract class ByCreation<PK extends Serializable, T, U> extends SequenceBaseEntity<PK>
-	implements
-		IdentifiableByCreatable<PK, T, U>
+	implements IdentifiableByCreatable<PK, T, U>
 {
 
 	/** The date and time when the entity that owns this entity was created. */