+++ src/main/java/io/github/astrapi69/entity/creatable/Creation.java
@@ -29,6 +29,7 @@ import java.io.Serializable;
 import javax.persistence.Entity;
 import javax.persistence.MappedSuperclass;
 
+import io.github.astrapi69.data.creatable.IdentifiableCreatable;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -53,8 +54,7 @@ import io.github.astrapi69.entity.base.SequenceBaseEntity;
 @NoArgsConstructor
 @SuperBuilder
 public abstract class Creation<PK extends Serializable, T> extends SequenceBaseEntity<PK>
-	implements
-		IdentifiableCreatable<PK, T>
+	implements IdentifiableCreatable<PK, T>
 {
 
 	/** The date and time when the entity that owns this entity was created. */