+++ src/main/java/io/github/astrapi69/entity/deletable/ByDeletion.java
@@ -29,6 +29,7 @@ import java.io.Serializable;
 import javax.persistence.Entity;
 import javax.persistence.MappedSuperclass;
 
+import io.github.astrapi69.data.deletable.IdentifiableByDeletable;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -69,8 +70,7 @@ import io.github.astrapi69.entity.base.SequenceBaseEntity;
 @NoArgsConstructor
 @SuperBuilder
 public abstract class ByDeletion<PK extends Serializable, T, U> extends SequenceBaseEntity<PK>
-	implements
-		IdentifiableByDeletable<PK, T, U>
+	implements IdentifiableByDeletable<PK, T, U>
 {
 
 	/** The date and time when the entity that owns this entity was deleted. */