+++ src/main/java/io/github/astrapi69/entity/deletable/Deletion.java
@@ -29,6 +29,7 @@ import java.io.Serializable;
 import javax.persistence.Entity;
 import javax.persistence.MappedSuperclass;
 
+import io.github.astrapi69.data.deletable.IdentifiableDeletable;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -67,8 +68,7 @@ import io.github.astrapi69.entity.base.SequenceBaseEntity;
 @NoArgsConstructor
 @SuperBuilder
 public abstract class Deletion<PK extends Serializable, T> extends SequenceBaseEntity<PK>
-	implements
-		IdentifiableDeletable<PK, T>
+	implements IdentifiableDeletable<PK, T>
 {
 	/** The date and time when the entity that owns this entity was deleted. */
 	private T deleted;