+++ src/main/java/io/github/astrapi69/entity/modifiable/ByLastModification.java
@@ -29,6 +29,7 @@ import java.io.Serializable;
 import javax.persistence.Entity;
 import javax.persistence.MappedSuperclass;
 
+import io.github.astrapi69.data.modifiable.IdentifiableByLastModified;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -58,8 +59,7 @@ import io.github.astrapi69.entity.base.SequenceBaseEntity;
 public abstract class ByLastModification<PK extends Serializable, T, U>
 	extends
 		SequenceBaseEntity<PK>
-	implements
-		IdentifiableByLastModified<PK, T, U>
+	implements IdentifiableByLastModified<PK, T, U>
 {
 
 	/** The date and time when the entity that owns this entity was modified. */