+++ src/main/java/io/github/astrapi69/entity/nameable/BasicNameEntity.java
@@ -28,6 +28,7 @@ import java.io.Serializable;
 
 import javax.persistence.MappedSuperclass;
 
+import io.github.astrapi69.data.nameable.IdentifiableNameable;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -49,8 +50,7 @@ import io.github.astrapi69.entity.base.SequenceBaseEntity;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 @SuperBuilder
 public abstract class BasicNameEntity<PK extends Serializable> extends SequenceBaseEntity<PK>
-	implements
-		IdentifiableNameable<PK>
+	implements IdentifiableNameable<PK>
 {
 
 	/** The name. */