+++ src/main/java/io/github/astrapi69/entity/nameable/activatable/ActivatableNameEntity.java
@@ -35,7 +35,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.experimental.FieldDefaults;
 import lombok.experimental.SuperBuilder;
-import io.github.astrapi69.entity.nameable.IdentifiableNameableActivatable;
+import io.github.astrapi69.data.nameable.IdentifiableNameableActivatable;
 import io.github.astrapi69.entity.nameable.NameEntity;
 
 /**