+++ src/main/java/io/github/astrapi69/entity/traceable/TraceableEntity.java
@@ -28,6 +28,7 @@ import java.io.Serializable;
 
 import javax.persistence.MappedSuperclass;
 
+import io.github.astrapi69.data.traceable.IdentifiableTraceable;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -57,8 +58,7 @@ import io.github.astrapi69.entity.modifiable.LastModification;
 @NoArgsConstructor
 @SuperBuilder
 public abstract class TraceableEntity<PK extends Serializable, T, U> extends SequenceBaseEntity<PK>
-	implements
-		IdentifiableTraceable<PK, T, U>
+	implements IdentifiableTraceable<PK, T, U>
 {
 
 	/** The date and time when the entity that owns this entity was created. */