+++ src/main/java/io/github/astrapi69/entity/traceable/versionable/VersionableTraceableEntity.java
@@ -35,7 +35,7 @@ import lombok.Setter;
 import lombok.experimental.SuperBuilder;
 import io.github.astrapi69.entity.creatable.Creation;
 import io.github.astrapi69.entity.deletable.Deletion;
-import io.github.astrapi69.entity.traceable.IdentifiableTraceableVersionable;
+import io.github.astrapi69.data.traceable.IdentifiableTraceableVersionable;
 import io.github.astrapi69.entity.traceable.TraceableEntity;
 
 /**