+++ src/main/java/io/github/astrapi69/entity/treeable/TreeUUIDEntity.java
@@ -31,6 +31,7 @@ import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.MappedSuperclass;
 
+import io.github.astrapi69.data.treeable.Treeable;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -50,8 +51,7 @@ import io.github.astrapi69.entity.uniqueable.UUIDEntity;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 @SuperBuilder
 public class TreeUUIDEntity<T, TR extends Treeable<T, TR>> extends UUIDEntity
-	implements
-		Treeable<T, TR>
+	implements Treeable<T, TR>
 {
 
 	/** The depth of this node. For the root depth would be 0. */