+++ src/main/java/io/github/astrapi69/entity/validatable/ValidatableEntity.java
@@ -29,6 +29,7 @@ import java.io.Serializable;
 import javax.persistence.Entity;
 import javax.persistence.MappedSuperclass;
 
+import io.github.astrapi69.data.validatable.IdentifiableValidatable;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -51,8 +52,7 @@ import io.github.astrapi69.entity.base.SequenceBaseEntity;
 @NoArgsConstructor
 @SuperBuilder
 public abstract class ValidatableEntity<PK extends Serializable, T> extends SequenceBaseEntity<PK>
-	implements
-		IdentifiableValidatable<PK, T>
+	implements IdentifiableValidatable<PK, T>
 {
 
 	/** The valid from date for validation. */