+++ src/main/java/io/github/astrapi69/entity/validatable/versionable/VersionableValidatableEntity.java
@@ -30,6 +30,7 @@ import javax.persistence.Entity;
 import javax.persistence.MappedSuperclass;
 import javax.persistence.Version;
 
+import io.github.astrapi69.data.validatable.versionable.IdentifiableValidatableVersionable;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -56,8 +57,7 @@ import io.github.astrapi69.entity.validatable.ValidatableEntity;
 public abstract class VersionableValidatableEntity<PK extends Serializable, T>
 	extends
 		ValidatableEntity<PK, T>
-	implements
-		IdentifiableValidatableVersionable<PK, T>
+	implements IdentifiableValidatableVersionable<PK, T>
 {
 
 	/**