+++ src/main/java/io/github/astrapi69/entity/verifiable/BasicVerifiableEntity.java
@@ -28,6 +28,7 @@ import java.io.Serializable;
 
 import javax.persistence.MappedSuperclass;
 
+import io.github.astrapi69.data.verifiable.Verifiable;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -35,7 +36,7 @@ import lombok.Setter;
 import lombok.experimental.FieldDefaults;
 import lombok.experimental.SuperBuilder;
 import io.github.astrapi69.entity.base.SequenceBaseEntity;
-import io.github.astrapi69.entity.identifiable.Identifiable;
+import io.github.astrapi69.data.identifiable.Identifiable;
 
 /**
  * The class {@link BasicVerifiableEntity} is a base entity that have to be signed and verified
@@ -51,8 +52,7 @@ import io.github.astrapi69.entity.identifiable.Identifiable;
 @SuperBuilder
 public abstract class BasicVerifiableEntity<PK extends Serializable> extends SequenceBaseEntity<PK>
 	implements
-		Identifiable<PK>,
-		Verifiable
+		Identifiable<PK>, Verifiable
 {
 
 	/** The signature. */