+++ CHANGELOG.md
@@ -4,7 +4,13 @@
 Version 2.4-SNAPSHOT
 -------------
 
+ADDED:
 
+- new dependency io.github.astrapi69:data-api in major version 3 that replace the dependency ent-api
+
+CHANGED:
+
+- removed dependency io.github.astrapi69:ent-api
 
 Version 2.3
 -------------